// src/api/model/userProfileDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserProfileDTO { 
    id?: number;
    name?: string;
    login?: string;
    role?: UserProfileDTO.RoleEnum;
    phone1?: string;
    phone2?: string;
    crea?: string;
    creaUf?: string;
    cpf?: string;
    creationDate?: string;
    enabled?: boolean;
    address?: string;
    addressNumber?: string;
    zipCode?: string;
    complement?: string;
    city?: string;
    state?: string;
}
export namespace UserProfileDTO {
    export const RoleEnum = {
        Admin: 'ADMIN',
        Gestor: 'GESTOR',
        Engenheiro: 'ENGENHEIRO',
        Fiscal: 'FISCAL',
        User: 'USER'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}